requirements.txt
python-telegram-bot==21.4
ccxt==4.3.91
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.9.2
